requirements.txt
streamlit
chromadb
openai
langchain
python-dotenv
pandas
numpy